!UPSTREAM x86/fpu: Adjust to handle misordered feature numbers
While new features are extended, some features are deprecated as no longer
useful. A new feature tends to have a higher number. But, its offset may be
lower to locate itself in the XSAVE buffer where a deprecated feature was.

This may help not to keep the uncompacted buffer size growing. But, the
kernel code has assumed that the highest-numbered feature has the highest
offset in the buffer.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>