KVM: x86: Advertise AMX-FP8 CPUID to user space
AMX-FP8 is a new set of instructions in the latest Intel platform
Diamond Rapids, for supporting FP8 datatype for AI usage.

The bit definition:
CPUID.(EAX=7,ECX=1):ECX[bit 3]

AMX-FP8 is on a new and sparse CPUID leaf and all bits on this leaf
have no truly kernel use case for now. Given that and to save space
for kernel feature bits, move this new leaf to KVM-only subleaf and
plus an x86_FEATURE definition for AMX-FP8 to direct it to the KVM
entry.

Advertise AMX-FP8 to KVM userspace. This is safe because there are
no new VMX controls or additional host enabling required for guests
to use this feature.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>