x86/microcode: Support for the uniform update scope
The microcode update process in the kernel currently operates under the
assumption of a core scope when identifying the control CPU and its
corresponding secondary CPUs.

With the introduction of the uniform scope concept, adjustments are
needed in the relevant code to accommodate this extended scope property.

Encapsulate topology_sibling_cpumask() with a new helper
ucode_get_scope_mask(), and then expand it to return a cpumask for these
additional scopes: package and system.

Given that the function’s call sites exclusively handle online CPUs, the
conversion to an online cpumask remains coherent with these sites.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org