x86/microcode: Clarify online enforcement behavior
The microcode update necessitates the presence of at least one primary
thread to oversee the update process for each defined scope. The existing
code mandates that every SMT primary thread remains online with the
current presumption of the core scope.

When the scope is extended to encompass the package or system, this
enforcement may be overly stringent, as more offline CPUs are allowable.
The question is whether this strict enforcement poses any practical
issues during real-world usage.

While the usefulness of the ‘nosmt’ approach is already acknowledged in
real-world scenarios, the practice of offline CPUs during microcode
update is not prevalent. So this online enforcement is presumed to be
benign for system administrators.

Considering this perspective, the existing code can be retained. Instead,
this aspect needs to be clarified. Add a comment to highlight this
consideration.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org