x86/msr-index: Define MSRs for the microcode uniform update
The microcode uniform feature streamlines the update process, expanding
the scope of CPUs to encompass the package or even the system-wide level,
as opposed to the current default core scope.

If MSR_IA32_ARCH_CAPABILITIES[16] is active, this CPU feature’s
capability and status are exported through these registers:

MSR_IA32_MCU_ENUMERATION:
    Bit 0: Set to indicate availability of the uniform feature
    Bit 1: Signifies that additional firmware configuration is required
    Bit 2: Set to confirm the completion of all necessary configurations
    Bits 8-15: Enumerates the uniform update scope

MSR_IA32_MCU_STATUS:
    Bit 0: Set if the last update resulted in a partial update
    Bit 1: Indicates an authentication failure occurred

Define the MSR indexes and specify the relevant bit positions and range
as preparatory.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: Hans de Goede <hdegoede@redhat.com>
Cc: Babu Moger <babu.moger@amd.com>
Cc: Reinette Chatre <reinette.chatre@intel.com>
Cc: Nikunj A Dadhania <nikunj@amd.com>
Cc: Kan Liang <kan.liang@linux.intel.com>
Cc: Breno Leitao <leitao@debian.org>
Cc: linux-kernel@vger.kernel.org
---
At the moment, this detail is not in the Intel SDM yet. But those can be
found in this publication:
    Runtime Microcode Update Technical Paper, June 2023:
    https://cdrdv2-public.intel.com/782715/356111-ucode-update.pdf

Hopefully, the above MSR description can help readers to get the idea
before looking up the specification.