x86/microcode/intel: Setup the uniform update scope
At the moment, the core code is equipped to determine the scope of CPUs
for the uniform update. The associated MSRs are defined to interpret CPU
capabilities and status. Also, the vendor-specific code can deactivate
the microcode update when needed.

With this foundation in place, the next step to prepare the uniform
update involves configuring the scope based on the CPU’s specifications.

Add an open code sequence under setup_uniform_update(), which evaluates
the CPU-exported bits and take measures to disable the update facility if
anything goes wrong. Then, invoke this setup function early to ensure the
udpate mechanism before any loading occurs.

Finally, set the scope and feature properties of struct microcode_ops if
the feature is ready to use.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org