x86/microcode/intel: Verify the uniform update status
Despite the fact that the uniform update leads to a partial update, the
revision is expected to be changed. The current validation code
considers such revision change alone is sufficient enough. Then, this
does not accurately interpret the error situation.

The feature is accompanied by MSR_IA32_MCU_STATUS, which exhibits
update information to the kernel. Inspect the MSR immediately after the
MSR_IA32_UCODE_WRITE write operation on the control CPUs.

Also, establish a handling policy to treat partial error or any unknown
conditions as system fatal. These situations are considerably with an
unstable microcode engine on the fly.

While the update is in the NMI context, raising a fatal should be
avoided. Record a fatal status during the rendezvous, and it will
eventually fire a fatal on the exit.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>
Cc: x86@kernel.org
Cc: linux-kernel@vger.kernel.org