x86/microcode: Add a fatal error state
Rasing panic() while rendezvous in NMI causes inter-dependency as
panic() broadcasts IPIs expecting CPUs exits a critical section soon.

Alternatively, record the fatal error in the per-CPU data and raise the
panic on the exit of rendezvous.

Introduce a new state indicating a fatal error, and record the new
state when a panic() is expected. Finally, when analyzing the
ucode_ctrl results, handle the fatal error as firing a panic().

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>