x86/microcode: Prepare a fatal error at boot time
The 'uniform' update is going to bring an error state indicating
'partial update'. Then, the running microcode has been updated but the
update didn't go through as intended.

The kernel needs to conservatively treat this error. A fatal error
during CPU bring-up is not directly translated into panic() as it is
not ready until the later boot stage.

Add a flag to capture the fatal condition from the early loading.
Then, check the flag and, when the flag is set, raise a panic right
before the loader setup.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>