kexec: Disable LASS before calling target kernel
On SRF we'd get a GPF when jumping to the target kernel because we're in
the upper half of the address space while the new kernel, being 1:1 mapped,
is quite likely in the lower half.  LASS protects from this type of calls.

While writing the cr4 register, take the chance to disable SMAP, SMEP and
CET to avoid similar issues.

Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>