KVM: x86: Advertise AVX10.2 CPUID to userspace
The initial, fully-featured version of Intel AVX10 will be enumerated as
AVX10.2, will include the new YMM-form of embedded rounding and Suppress
All Exceptions (SAE), the new enumeration scheme, and several new sets
of instructions.

Since the new AVX10.2 instructions can't be intercepted, and AVX10 state
enumeration in CPUID leaf D and enabling in XCR0 register are identical
to AVX512 state regardless of the supported vector lengths, advertise
AVX10.2 if it's supported in hardware.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>