crypto: qat - register supported algorithms with crypto framework
Support for cipher or aead algorithms differs across generations.
Check device capability before registering an algorithm with the crypto
framework.
If the algorithm is not supported by the QAT hardware, it will fallback to
software implementation.
AES-CTR-192 key size is deprecated from GEN6 HW devices. Hence software
fallback is implemented to support this key size in the self test.

Signed-off-by: Karthikeyan Gopal <karthikeyan.gopal@intel.com>