KVM: x86/pmu: Let the host handle Intel PT interrupts in system mode
Intel PT is not exposed to the guest in PT system mode, and the host
is supposed to be allowed to create system-wide PT events.  In non-root
mode, PT-induced PMIs are delivered to the KVM GUEST PMI handler, and
it needs to be handled by the host perf, instead injecting into the
guest.

Signed-off-by: Zhenyu Wang <zhenyuw@linux.intel.com>
Signed-off-by: Zide Chen <zide.chen@intel.com>