x86/microcode/intel: Prepare for staging microcode
When staging is initiated, operations are conducted through an MMIO
interface. Each interface is independently discoverable via
IA32_MCU_STAGING_MBOX_ADDR MSR and linked 4 to a set of dword-size
registers.

Key hardware specifications necessitate two primary tasks ensuring
staging for every exposed MMIO space and validating the dword alignment
of the microcode image.

Upon receiving a microcode image, check first for dword alignment.
Subsequently, it iterates through each CPU to locate staging addresses
and conducts staging for every unique MMIO interface. Although this
process may be tedious, it is acceptable as staging occurs in a
non-critical path.

Signed-off-by: Chang S. Bae chang.seok.bae@intel.com
---
This serialized staging is not necessary but simple. Staging is performed
on a slow path. But, the latency is observable via 'time echo 1 >
/sys/devices/system/cpu/microcode/reload'. If this serialization becomes
really objected, then it could be parallelized.