x86/microcode/intel_staging: Support data transfer
The staging architecture features a narrowed interface for data transfer.
Instead of allocating MMIO space based on data chunk size, it utilizes
two data registers: one for reading and one for writing, enforcing the
serialization of read and write operations. Additionally, it defines a
data format.

To facilitate data transfer, implement helpers according to the specified
format for writing or reading the staging data.

Signed-off-by: Chang S. Bae <chang.seok.bae@intel.com>