crypto: qat - enable RAS support for GEN6 devices
Reliability, Accessibility, Serviceability (RAS) enables the Gen6 Kernel
driver to take appropriate actions depending on the severity of the error.
The actions range from a simple notification to the user to a function
level reset, The errors are read in error source registers referenced
ERRSOU0 to ERRSOU3. Each occurrence of an error is incrementing a counter
specific to the error type.

This commit adds:
- The infrastructure to handle the errors from ERRSOU 0 - 3.
- Enabled the RAS error counters in sysfs for GEN6 HW

The counters added are:
- errors_correctable - Hardware correctable errors include the
  error conditions where the system can recover without any loss of
  information.
- errors_nonfatal - This category of errors includes the errors that
  can be isolated to one or more requests in-flight.
- errors_fatal - This category of errors includes the errors that
  cannot be contained and renders the hardware unreliable.
  QAT HW enters in DeviceHALT state and an FLR is required to return to a
  reliable mode of operation.

Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>