EMBARGOED selftests/sgx: 128 byte SSL interactions
Since the SHA384 opt-in mode has a longer key the relevant data
structures and the digest update function need to be updated. The
reserved space in the data structures needs to make them 128 byte long
instead of the current 64 bytes.

Add SHA384 compatible data structures where relevant and make sure
they're only used when the SHA384 mode is enabled.

Make the mrenclave_update() rely on the hash_type for how much data is
hashed so the SHA256 legacy mode and the opt-in SHA384 mode can both
work.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>