KVM: x86/pmu: Implement guest Arch LBR contex switch
In mediated vPMU, if guest LBR is enabled, LBR MSRs are passthrough,
and the guest owns the LBR related resources in non-root mode, while
host owned them in root mode.

LBR context needs to be switched in VMX transitions.  To save guest LBR
context, MSRs are cleared to avoid leaking guest LBRs into host.

Guest Arch LBR is enabled only if XSAVE Arch LBR feature is available
on the host, and xsaves/xrstors are used for fast MSR accesses.

"if (lbr_desc->msr_passthrough)" is the only criterion to determine
whether or not the guest context needs to be saved or restored, which
has a few implications:

- Power on default value is zero: the KVM LBR state area
  (lbr_desc->state) is initialized to zero, and before guest Arch LBR
  is enabled, LBR triplet MSRs are intercepted and guest read as all 0.

- When guest Arch LBR is disabled, Arch LBR MSRs are intercepted, and
  the writes from guest are cached in KVM and will not be restored to
  the guest context until next time guest Arch LBR is enabled.

Signed-off-by: Xiong Zhang <xiong.y.zhang@linux.intel.com>
Signed-off-by: Zide Chen <zide.chen@intel.com>