KVM: x86/pmu: Emulate Arch LBR MSRs in mediated vPMU
When the guest Arch LBR is enabled, the related MSRs are passthrough
and need to be switched between the host counterparts at VM entry/exit.
KVM needs to emulate them for accesses

When the guest LBR is disabled, indicating that the guest is not using
LBR, KVM does not need to switch LBR context at VM entry/exit, and
intercepts and emulates these MSRs.  Also accesses from userspace need
to be emulated as well.

Signed-off-by: Xiong Zhang <xiong.y.zhang@linux.intel.com>
Signed-off-by: Zide Chen <zide.chen@intel.com>