perf/x86/intel/cstate: Apply multi-die/package workaround for CLX-AP only
Old commit cb63ba0f670d ("perf/x86/intel/cstate: Support multi-die/package")
tried to workaround issue on CLX-AP that punit is per-die instead of package
level. It has no problem on system as normally there's only single die.

But for DMR, CBB is die. Old change would make package cstate on DMR to be
marked as per-die which is not correct. It still keeps as package level.

This one tries to apply old workaround to CLX-AP only which is by introducing
a new quirk for SKL type, but not propagate to other platform.

[Rebase: resolve against Kan's pmu scope change]

Cc: Rui Zhang <rui.zhang@intel.com>
Cc: Artem Bityutskiy <artem.bityutskiy@intel.com>
Cc: Wendy Wang <wendy.wang@intel.com>
Cc: Kan Liang <kan.liang@linux.intel.com>
Signed-off-by: Zhenyu Wang <zhenyuw@linux.intel.com>