EDAC/{skx_common,imh}: Add EDAC driver for Intel Diamond Rapids server processors
Intel Diamond Rapids CPUs are built on two types of dies: Integrated Memory
Hubs (IMH) and Core Building Blocks (CBB). The memory controller in this
type of CPU is integrated within the IMH to provide a memory stack to the
processor. Create a new driver for this IMH-based memory controller rather
than applying additional patches to i10nm_edac.c for the following reasons:

1) The memory controllers are not presented as PCI devices; instead,
   the detection and all their registers have been transitioned to
   MMIO-based memory spaces.

2) Validation processes are costly. Modifications to i10nm_edac would
   require extensive validation checks against multiple platforms,
   including Ice Lake, Sapphire Rapids, Emerald Rapids, Granite Rapids,
   Sierra Forest, and Grand Ridge.

Tested-by: Yi Lai <yi1.lai@intel.com>
Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>