hwmon: (coretemp) Create package-scoped platform device
Currently, the coretemp driver creates a platform device for each die.
This approach works well on single-die systems and existing
multi-die system Cascade Lake-AP (CLX-AP). However, it will encounter
issues on new multi-die systems. For example, on a Diamond Rapids (DMR)
system with four dies in a package, it will create four coretemp
devices, but they will have identical information because the thermal
MSRs of DMR systems are package-scoped and shared by all the dies within
a package. It would be ideal to avoid the duplicate devices and create
only one device in this case.

There may be two options to address this issue:

Option 1: Treat DMR as an exception, and create die-scoped coretemp
device for all other platforms, including CLX-AP.

Option 2: Treat CLX-AP as an exception, and create package-scoped
coretemp device for all other platforms, including DMR.

Considering that the majority of platforms use package-scoped thermal
MSRs by nature, it is more practical to adopt option 2 as the solution
for this issue. After the patch:

- For CLX-AP system, a coretemp device is created per die.
- For all other systems, a coretemp device is created per package.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>