vfio/qat: add support for intel QAT 6xxx virtual functions
Add support for QAT 6xxx devices in the QAT vfio-pci variant driver.
As the differences between QAT 6xxx and QAT 4xxx VFs are abstracted
by the QAT PF driver, there are no changes required apart from
the addition of the VF ID on the probe table.

The module description has been changed to indicate that this module is
generic to all QAT generation

Signed-off-by: Malgorzata Mielnik <malgorzata.mielnik@intel.com>