iommu/vt-d: HPT capability enumeration and configuration
The HPTS bit in the extended capability register indicates if hardware
supports HPT. This capability only applies to the Scalable Mode. If the
hardware advertizes the HPT feature, enable it by setting the HPTE bit
in the Scalable Mode context entry.

The HPT root table address together with other controls are encoded in
the scalable mode pasid entry. This sets up those fields as well.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Jingqi Liu <Jingqi.liu@intel.com>
Signed-off-by: Zeng Guang <guang.zeng@intel.com>
Signed-off-by: Yang Zhong <yang.zhong@linux.intel.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>#refactor the setup/replace helpers, only flush hpt cache in pasid replace, no in pasid setup