crypto: qat - add live migration support for GEN6 devices
The current implementation of the QAT live migration feature is exclusive
to GEN4 devices and resides within GEN4 specific files. However, the
underlying mechanisms, including the associated CSRs and offsets, are
shared between GEN4 and GEN6 devices.
Added access to the GEN4 live migration functionality for GEN6 devices
through the abstraction layer.
Point the bank save and restore functions in the GEN6 driver to the
common code.

Signed-off-by: Malgorzata Mielnik <malgorzata.mielnik@intel.com>