KVM: x86/pmu: Handle fixed counter enumeration CPUID(0xA).ECX
Perfmon v5 introduces CPUID(0xA).ECX to enumerate fixed counters.

In detail, SDM says "Software is recommended to use the following logic
to check if a Fixed Counter is supported on a given processor:
	FxCtr[i]_is_supported := ECX[i] || (EDX[4:0] > i); "

So the fixed counter i is supported either it's indicated in ECX[i] == 1
or EDX[4:0] > i, in other words, ECX should be subset of or equal to the
fixed counter indicated in EDX[4:0].

Thus support fixed counter enumeration in CPUID(0xA).ECX and check if it
is correctly configured by user space.

Signed-off-by: Yang Weijiang <weijiang.yang@intel.com>