EMBARGOED x86/sgx: Allow SGX256 enabling
The Secure Enclave Control Structure (SECS) includes an attributes
field. These attributes ensure consistency and transparency around
specific enclave features. Some of these attributes are passed from
userspace during the SGX_IOC_ENCLAVE_CREATE ioctl. In the case of SHA384
support, a new attribute was introduced starting with Diamond Rapids.
This attribute indicates an opt-in to use a lengthened hash,
transitioning from the legacy SHA256 to SHA384.

The kernel keeps track of both privileged and unprivileged attributes.
The unprivileged attributes are by default allowed to be opt-in when
creating the enclave but the privileged ones need userspace to
explicitly opt-in. SGX256 enhances the enclave security by using
stronger keys thus can be treated as an unprivileged attribute.

Some fields in the SECS structure are reserved, and SGX_ATTR_SHA_384 is
taking the place of a reserved field. This bit should be excluded from
the reserved mask thus allowing the enabling of SHA384.

Changelog v4:
- Create this patch and move SHA384 attribute code here.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>