EMBARGOED KVM: x86: Extend IA32_SGX_LEPUBKEYHASH MSRs for SGX256
If 256-bit SGX extensions are available, the array size of
IA32_SGXLEPUBKEYHASH will be extended to six. Add two new MSRs
IA32_SGX_LEPUBKEYHASH[4,5] and inject #GP when the availability of
reads/writes to the MSRs is violated.

Also outfit sgx_virt_einit() with the ability to opt-in into SHA384 by
removing the previously hardcoded SHA256 algorithm for hashing.

Changelog v4:
- Move virt.c code from host patch since it's only used in kvm.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>