KVM: pmu/x86: Fix guest Intel PT enable issue
If IA32_VMX_MISC[14] is 0, Intel PT is not allowed to be used in VMX
operation (including in root mode), which means Intel PT on the guest
is not enabled if the vCPU is post-VMXON (pt_pmu.vmx == true),
regardless whether a nested guest is created or not.

This becomes more vulnerable after commit 9a798b1337af ("KVM: Register
cpuhp and syscore callbacks when enabling hardware"), which executes
VMXON on host CPUs when enabling virtualization in hardware during KVM
module initialization, other than when creating a guest VM.

This is fixed by exposing this bit to the guest.  This patch cleans up
previous fixes on emulating writes to IA32_RTIT_CTL.  In VMX operation,
writes to IA32_RTIT_CTL are not allowed if IA32_VMX_MISC[14] is 0.
vmx_pt_mode_is_host_guest() checks the PT host guest mode, which is not
true only if IA32_VMX_MISC[14] is cleared.  This is sufficient to
justify the Intel PT permission in VMX operation, and the check of
vmx->nested.vmxon needs to be removed.

Fixes: ee85dec2fe9c ("KVM: x86: Disable Intel PT when VMXON in L1 guest")
Signed-off-by: Zide Chen <zide.chen@intel.com>
Reviewed-by: Zhenyu Wang <zhenyuw@linux.intel.com>