iommu/vt-d: Add intel_iommu=ignore for test and debug
The intel_iommu=ignore kernel boot option instructs the Intel IOMMU driver
to bypass the IOMMU hardware and leave it in its current state. This is
useful for test cases where VT-d is already configured by software prior
to OS boot. The test solely relies on the OS driver for its execution.

This behavior is largely equivalent to not compiling the Intel IOMMU driver
(CONFIG_INTEL_IOMMU=n), with the exception that interrupt remapping remains
functional.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>