crypto: qat - add mechanism to reset based on error threshold
Treat a cascade of uncorrectable errors as a fatal error.
The cascade of uncorrectable errors is detected if the uncorrectable error
counter increments by more than a configurable threshold within a
time interval. The threshold and the time interval should be configurable.

Add a timer thread that checks whether the number of uncorrectable
errors is above the threshold. An FLR is issued if the number of
uncorrectable errors is more than the threshold within a time interval.
Added adf_ras.c file to have the common implementation of the non-fatal
error timer thread.

The below sysfs entries are added:
- errors_nonfatal_timer: This sysfs entry is used to configure the
  time interval in milliseconds to check the count of non-fatal
  errors in the device.
- errors_nonfatal_threshold: This sysfs entry is used to configure the
  threshold for non-fatal errors within the time interval configured in
  errors_nonfatal_timer. When the number of errors exceeds the threshold,
  an FLR is issued.

Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>