x86/{msr,mce}: Enumerate the bit-fix filter reset bit
Intel processors with the bit MCG_TES_P of IA32_MCG_CAP set support error
reporting based on the 'threshold-based error status'. The threshold is
chosen by Intel based on various factors. This status provides an
indication of the health of the associated hardware component (e.g., the
CPU cache).

  Green  : the number of correctable errors <= threshold.
  Yellow : the number of correctable errors  > threshold.

The yellow indication is tied to a BFF (Bit-Fix Filter) buffer that keeps
track of the hardware component that have experienced correctable errors.
However, it doesn't have any relationship to time since it simply
accumulates corrected errors as they are observed. Hence, the problem is
that it's impossible to know whether the yellow indication is set due to a
rush of corrected errors that happened in a short time (e.g., within 10
minutes) or simply an accumulation of corrected errors over a long term
(e.g., the past 5 years) that the system has been up.

Newer Intel systems provide the capability to reset the bit-fix filter.
Enumerate this feature as a preparation for the upcoming patch that resets
the bit-fix filter. That will alert users if the yellow status repeats
frequently, indicating that there are enough real defects to overflow the
bit-fix filter. In such cases, the system will need to be scheduled for
servicing.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>