x86/mce: Reset the bit-fix filter when it overflows
Intel processors with MCG_TES_P capability internally use the bit-fix
filter buffer to tracks corrected errors of some hardware components.
If the bit-fix filter overflows, a 'yellow' flag will be set in the
MCi_STATUS[54:53] to indicate the corrected errors of the associated
hardware component exceed the pre-defined threshold.

However, it is unknown in which of the following cases the bit-fix
overflow occurs:

  a) A long-term accumulation of corrected errors (e.g., in the past 5 years).
  b) A transient storm of corrected errors (e.g., in the last 10 minutes).
  c) A lasting storm of corrected errors that repeatedly overflow the bit-fix filter.

For cases a) and b), it may be necessary to observe the system closely,
but no immediate action may be required.

For case c), it indicates that there are enough real defects to overflow
the bit-fix filter, which might lead to uncorrected errors soon. Therefore,
the system needs to be scheduled for servicing.

Newer Intel systems provide the capability to reset the bit-fix filter.
Reset the bit-fix filter when it overflows and check the overflow frequency
to determine if case c) above is occurring.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>