KVM: VMX: Implement guest_can_use_arch_lbr()
During the time to initalize KVM and setup guest VM, there are
multiple places that need to check if Arch LBR is available and can be
exposed to the guest.  Implement this API to reduce redundant code,
and to make the code more readable.

Any platforms support Arch LBR are supposed to support XSAVE/XRSTORS
as well, thus we make the guest Arch LBR feature explicitly depend on
it, and we can use xsave/xrstors for LBR context switching.

Signed-off-by: Zide Chen <zide.chen@intel.com>