KVM: x86/pmu: Include legacy LBR MSRs logic in legacy platforms only
Legacy LBR is available on legacy platforms that architectural LBR is
not available.  Clean up the common code that may handle both legacy
LBR and Arch LBR, to make sure legacy MSRs are not accessed in case
Arch LBR is available.

The sanity check against lbr_desc->msr_passthroug in intel_pmu_refresh()
is to make sure PMU refresh is not called after first KVM_RUN, adjusted
the comment to reflect that it applies to either legacy or mediated
vPMU.

Signed-off-by: Xiong Zhang <xiong.y.zhang@linux.intel.com>
Signed-off-by: Zide Chen <zide.chen@intel.com>