EMBARGOED x86/sgx: Add new MSR's to support SHA384 hashes
The EINIT instruction requires an RSA-3072 signature of the enclave
measurement. For EINIT to function properly, the public key used for
signing the measurement must be hashed and written to the
IA32_SGXLEPUBKEYHASH MSR. This is part of the SGX Launch Control
feature. The hash algorithm used for hashing was previously SHA256, but
since Diamond Rapids, SGX has been extended to work with the more secure
SHA384. To accommodate this, two additional MSRs -
MSR_IA32_SGXLEPUBKEYHASH4 and MSR_IA32_SGXLEPUBKEYHASH5 - have been
added, enabling the MSRs to store the full SHA384 hash, which is 48
bytes long. If SHA384 is available but SHA256 is used the new MSRs need
to be zeroed out.

Furthermore, the IA_32_SGXLECONFIG MSR has been introduced. This new MSR
should be used to indicate the Launch Control key strength. At this
time, only one bit of this MSR is relevant, with the rest being
reserved. If bit 0 is turned off, the legacy mode will be used, and only
the four existing MSRs will be used to store the 32-byte long hash. If
the bit is turned on, the new mode will be activated, and the two newly
added MSRs will be used to store the lengthened hash.

SGX256 also extends SECS and SIGSTRUCT SGX structures by repurposing
their reserved spaces: Two new 'mrsigner384' and 'mrenclave384' fields
are added to SECS; One new 'sighashtype' is added to SIGSTRUCT and the
'mrenclave' is extended from 256-bits to 384-bits. Extend the kernel SGX
architectural structures to support them.

Changelog v5:
- Zero out MSR_IA32_SGXLEPUBKEYHASH4/5 if the key is 32 bytes long.

Changelog v4:
- Split off software changes to a separate patch.
- Update two patch message paragraphs so they're more organized and
  remove inaccuracies.
- Move all code needed specifically for sha384 here.

Changelog v3:
- Add separate sighash structure that hold hash information along with
  mrsigner used to updated MSRs.
- Add preset structures to copy into dynamically allocated ones during
  MSR updated flow.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>