crypto: qat - add support for rate limiting VF2PF message types
Add support for ADF_VF2PF_MSGTYPE_CIR_REQ and ADF_VF2PF_MSGTYPE_PIR_REQ
VF2PF message types.
These message types are used by applications using a VF to query the rate
limiting SLA assigned for the ring pair(RP) passed as a parameter in the
message data.
The handler will return the max SLA value to indicate that it is not
supported in GEN6 and GEN4 drivers.

Co-developed-by: George Abraham P <george.abraham.p@intel.com>
Signed-off-by: George Abraham P <george.abraham.p@intel.com>
Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>