KVM: x86: Set PT VMCS controls iff Intel PT is enabled in the guest
Set the Intel PT related VMCS controls only when Intel PT is actually
exposed to the guest.  This is to avoid extra overhead to load and
clear IA32_RTIT_CTL during VMX transitions in case the KVM is in
PT_MODE_HOST_GUEST mode, but it doesn't expose Intel PT to all guests.

Signed-off-by: Zide Chen <zide.chen@intel.com>