KVM: x86/pmu: Refactor Intel PT guest context switch code
Clear IA32_RTIT_CTL.TraceEn before VM entry, regardless whether or not
guest PT tracing is currently enabled or not.  This is to meet the
VM entry requirement as stated by SDM:

  If the logical processor is operating with Intel PT enabled (if
  IA32_RTIT_CTL.TraceEn = 1) at the time of VM entry, the "load
  IA32_RTIT_CTL" VM-entry control must be 0.

Notes: Intel PT PMU has PERF_PMU_CAP_EXCLUSIVE and it's context is not
saved and restored when it's being scheduled out/in.  Contrary to other
PMU features that rely on host perf to do that, KVM needs to save and
restore host Intel PT context in VMX transitions.

Rename APIs to match the naming stype in pmu_intel.c:
- pt_guest_enter() -> intel_pmu_load_guest_pt()
- pt_guest_exit() -> intel_pmu_put_guest_pt()

Signed-off-by: Zide Chen <zide.chen@intel.com>