perf/x86/intel: Support PERF_PMU_CAP_MEDIATED_VPMU for Intel PT PMU
Apply the PERF_PMU_CAP_MEDIATED_VPMU for Intel PT PMU.  When Intel PT
virtualization is enabled in KVM, this allows perf core to schedule out
all exclude_guest Intel PT events in perf_guest_enter().  This is needed
to allow the guest has exclusive access to Intel PT hardware after guest
Intel PT context is loaded.

The host is not allowed to create an !exclude_guest perf event if a PMU
with PERF_PMU_CAP_MEDIATED_VPMU capability and the hypervisor actually
enables it;  Meanwhile, an !exclude_guest event should be allowed in the
host if the hypervisor doesn't enable PERF_PMU_CAP_MEDIATED_VPMU on the
PMU that is associated with the perf event.

If a perf event is associated with a core PMU, perf core can evaluate
nr_mediated_pmu_vms to learn whether PERF_PMU_CAP_MEDIATED_VPMU is
enabled or not by the hypervisor.

For non-core PMU, PERF_PMU_CAP_MEDIATED_DISABLED_VPMU is introduced
to serve the same purpose, and this allows the host to be able to
create system-wide (w/o exclude_guest attribute) perf events if Intel
PT virtualization is not enabled.

Signed-off-by: Zide Chen <zide.chen@intel.com>