crypto: qat - enable telemetry for GEN6 devices
Add support for telemetry for GEN6 devices.
This feature is enabled on demand by writing a value between 1-4 to the
telemetry debugfs 'control' file.
Writing 1 would mean the user wants the current value only.
Whereas writing values 2-4 would give the current value and the max, min &
avg of the values over the provided number of samples.

This includes:
- Telemetry layout definition for GEN6 device.
- Initialize 'adf_tl_hw_data' for GEN6 device which is required for
  monitoring telemetry.

Co-developed-by: George Abraham P <george.abraham.p@intel.com>
Signed-off-by: George Abraham P <george.abraham.p@intel.com>
Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>