crypto: qat - qat_kpt sysfs folder unavailable
The qat_kpt sysfs folder is removed when the qat_6xxx module is unloaded.
The qat_kpt sysfs folder should not be removed when the device is
brought down as it is required to configure KPT in device-down state.
The KPT sysfs entries will be automatically removed when the qat_6xxx
module is unloaded.

Signed-off-by: Suman Kumar Chakraborty <suman.kumar.chakraborty@intel.com>