crypto: qat: Fix VFs re-enabling in recovery flow
Introducing of the configuration ADF_GENERAL_SEC persistence resulted in
having the entry ADF_SRIOV_ENABLED not removed even when the SRIOV was
disabled. This was mitigated during the UACCE implementation by adding
a function which would remove the entry when the SRIOV was disabled.

Unfortunately, the function was called also during the recovery flow,
which resulted in the VFs not being re-enabled after a device reset.

This patch fixes the issue by adding a check to not remove the entry
when the ADF_STATUS_RESTARTING bit is set.

Signed-off-by: Damian Muszynski <damian.muszynski@intel.com>