intel_idle: Support user-defined C-state table via kernel parameter
In certain scenarios, users may need to experiment with C-states that
are not present in the _CST or the built-in custom C-state table, or
they may need to modify the target residency or exit latency values of
existing C-states.
The current approach requires rebuilding the kernel with an experimental
patch, which is not flexible and can be challenging for users who lack
kernel build experience.

To address this, introduce a new module parameter "intel_idle.table=".
This parameter allows the intel_idle driver to ignore the internal
custom C-state table and build a new C-state table based on the input
provided via the kernel parameter. All other existing module parameters
that operate the C-state table will continue to function as before,
except that they will now operate on the new user-defined C-state table
rather than the internal built-in C-state table.

The input for "intel_idle.table=" is a space-separated string, where
each section of the string describes a C-state in the following format:

 "cstate_name:mwait_hint_value:exit_latency_us:target_residency_us"

For example, using below kernel parameter
 intel_idle.table="C1:0x0:1:1 C1E:0x1:2:2 C6:0x20:50:150 C10:0x60:100:300"
brings following C-state table

 /sys/devices/system/cpu/cpu0/cpuidle/state0/name:POLL
 /sys/devices/system/cpu/cpu0/cpuidle/state0/desc:CPUIDLE CORE POLL IDLE
 /sys/devices/system/cpu/cpu0/cpuidle/state0/latency:0
 /sys/devices/system/cpu/cpu0/cpuidle/state0/residency:0

 /sys/devices/system/cpu/cpu0/cpuidle/state1/name:C1
 /sys/devices/system/cpu/cpu0/cpuidle/state1/desc:MWAIT 0x0
 /sys/devices/system/cpu/cpu0/cpuidle/state1/latency:1
 /sys/devices/system/cpu/cpu0/cpuidle/state1/residency:1

 /sys/devices/system/cpu/cpu0/cpuidle/state2/name:C1E
 /sys/devices/system/cpu/cpu0/cpuidle/state2/desc:MWAIT 0x1
 /sys/devices/system/cpu/cpu0/cpuidle/state2/latency:2
 /sys/devices/system/cpu/cpu0/cpuidle/state2/residency:2

 /sys/devices/system/cpu/cpu0/cpuidle/state3/name:C6
 /sys/devices/system/cpu/cpu0/cpuidle/state3/desc:MWAIT 0x20
 /sys/devices/system/cpu/cpu0/cpuidle/state3/latency:50
 /sys/devices/system/cpu/cpu0/cpuidle/state3/residency:150

 /sys/devices/system/cpu/cpu0/cpuidle/state4/name:C10
 /sys/devices/system/cpu/cpu0/cpuidle/state4/desc:MWAIT 0x60
 /sys/devices/system/cpu/cpu0/cpuidle/state4/latency:100
 /sys/devices/system/cpu/cpu0/cpuidle/state4/residency:300

Signed-off-by: Zhang Rui <rui.zhang@intel.com>