Revert "platform/x86/intel/ifs: Fix SBAF title underline length"
This reverts commit 1e701372d7ac1939d5f8a1dc8172de00192394a8.

Structural Based Functional Test at Field (SBAF) test was expected to be
supported in Granite Rapids(GNR) cpus. However, due to recent roadmap
changes, SBAF test support will not be included in GNR and is unlikely
to be included in future cpus.

Given this, remove the SBAF test from the In-Field Scan (IFS) driver.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>