Revert "platform/x86/intel/ifs: Add SBAF test support"
This reverts commit 3c4d06bd6e3713235fba5aa5eed9d1898239ec1f.

Structural Based Functional Test at Field (SBAF) test was expected to be
supported in Granite Rapids(GNR) cpus. However, due to recent roadmap
changes, SBAF test support will not be included in GNR and is unlikely
to be included in future cpus.

Given this, remove the SBAF test from the In-Field Scan (IFS) driver.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>