Revert "platform/x86/intel/ifs: Add SBAF test image loading support"
This reverts commit 0a3e4e94d137daacd5ec092365080eed847f8f01.

Structural Based Functional Test at Field (SBAF) test was expected to be
supported in Granite Rapids(GNR) cpus. However, due to recent roadmap
changes, SBAF test support will not be included in GNR and is unlikely
to be included in future cpus.

Given this, remove the SBAF test from the In-Field Scan (IFS) driver.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>