drivers/platform/x86/ifs: Compute CPUs participating in IFS
Currently, the driver selects the CPUs to run the test together by
selecting the smt_mask of the CPU under test. This information is also
provided by the INTEGRITY_MSR's bits 20:16. In future CPUs, it could extend
to module scoped CPUs. Use this enumerated value to compute the CPUs
list dynamically.

This solves 2 problems:

- When a thread of offline, it's not possible to find out from the smt_mask,
since that CPU is removed from the smt_mask.
- The groups of CPUs that must be run together may not be covered by any of
the pre built topology masks. Change to building a custom mask.

This topology doesn't change, even if a CPU is offline since it's built from
the x2apic id's. This could be built one time at driver load time for all
CPUs, but it would be a per_cpu mask and would be taking space on a large
system. Compared to the latency of scheduling the test, building this mask
shouldn't not be a concern.

When one thread is offline, new log looks as below:

[ 1185.579751] misc intel_ifs_0: Some of the CPU(s) 96 are offline: SCAN operation did not start.

Build the mask based on the TID_SHIFT enumerated in the INTEGRITY_MSR and
the x2apic ID of the CPU under test. Use stop_machine_cpuslocked() always.

Signed-off-by: Jithu Joseph <jithu.joseph@intel.com>